+++ src/Api/QuestApi.cs
@@ -1,7 +1,12 @@
 ﻿using QuestFramework.Events;
+using QuestFramework.Framework;
 using QuestFramework.Framework.Events;
 using StardewModdingAPI;
 using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
 namespace QuestFramework.Api
 {
@@ -54,9 +59,7 @@ namespace QuestFramework.Api
                 modUid: manifest.UniqueID,
                 questManager: this.mod.QuestManager,
                 questOfferManager: this.mod.QuestOfferManager,
-                hookManager: this.mod.HookManager,
-                monitor: this.monitor
-            );
+                hookManager: this.mod.HookManager);
         }
 
         public void ForceRefresh()