+++ src/Quests/ITriggerLoader.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace QuestFramework.Quests
 {
-    interface ITriggerLoader
+    public interface ITriggerLoader
     {
         void LoadTrigger(string triggerData);
     }